GIT stuff & Editor var mapping
added gitattributes file and a revisiontagger to reflect a
SVN-a-like revision. That way the last commit hash will be added
(hopefully) to the VERSION variable which is useful when trouble
shooting. Note: That will work only when addrevision is in your
path and git was configured with smudge & clean

Introduced a variable mapping for the Editor which means:
a) the VMX settings are variables instead of hardcoded ones
b) the displayed description of a VMX variable is a variable now,too